refactor(io): remove testing reference
Save binary size
