+++ internal/io/mock/test_source.go
@@ -22,6 +22,7 @@ import (
 	"time"
 
 	"github.com/lf-edge/ekuiper/internal/converter"
+	mockContext "github.com/lf-edge/ekuiper/internal/io/mock/context"
 	"github.com/lf-edge/ekuiper/internal/topo/context"
 	"github.com/lf-edge/ekuiper/pkg/api"
 	"github.com/lf-edge/ekuiper/pkg/ast"
@@ -35,7 +36,7 @@ func TestSourceOpen(r api.Source, exp []api.SourceTuple, t *testing.T) {
 		count.Store(1)
 		c = 0
 	}
-	ctx, cancel := NewMockContext(fmt.Sprintf("rule%d", c), "op1").WithCancel()
+	ctx, cancel := mockContext.NewMockContext(fmt.Sprintf("rule%d", c), "op1").WithCancel()
 	cv, _ := converter.GetOrCreateConverter(&ast.Options{FORMAT: "json"})
 	ctx = context.WithValue(ctx.(*context.DefaultContext), context.DecodeKey, cv)
 	count.Store(c.(int) + 1)
